requirements.txt
affine==2.4.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
blinker==1.9.0
bokeh==3.6.3
boto3==1.37.1
botocore==1.37.1
branca==0.8.1
Brotli==1.1.0
cached-property==1.5.2
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
cytoolz==1.0.1
dash==2.18.2
dash_ag_grid==31.2.0
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash_html_components==2.0.0
dash_leaflet==1.0.15
dash-table==5.0.0
dask==2025.2.0
debugpy==1.8.12
decorator==5.2.1
distributed==2025.2.0
editables==0.5
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.0.3 # pinned by Dash 2.18
Flask-Caching==2.3.1
folium==0.19.4
fonttools==4.56.0
fsspec==2025.2.0
geobuf==2.0.0
geopandas==1.0.1
gunicorn==23.0.0
h2==4.2.0
h5netcdf==1.5.0
h5py==3.13.0
hatchling==1.27.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
locket==1.0.0
lz4==4.3.3
mapclassify==2.8.1
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
msgpack==1.1.0
munkres==1.1.4
narwhals==1.27.1
nest_asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0.1
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
prompt_toolkit==3.0.50
protobuf==5.28.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
Pygments==2.19.1
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.1
rasterio==1.4.3
requests==2.32.3
retrying==1.3.4
s3transfer==0.11.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
shapely==2.0.7
six==1.17.0
snuggs==1.4.7
sortedcontainers==2.4.0
stack_data==0.6.3
tblib==3.0.0
threadpoolctl==3.5.0
tomli==2.2.1
toolz==1.0.0
tornado==6.4.2
traitlets==5.14.3
trove-classifiers==2025.2.18.16
typing_extensions==4.12.2
tzdata==2025.1
unicodedata2==16.0.0
urllib3==2.3.0
wcwidth==0.2.13
waitress==3.0.2
Werkzeug==3.0.6 # pinned by Dash 2.18
wheel==0.45.1
xarray==2025.1.2
xyzservices==2025.1.0
zict==3.0.0
zipp==3.21.0
zstandard==0.19.0
